implemented some route in rest api and cleaned up entity mapping && added rule to authorize accessing trending route without login

+++ storyAPI/src/main/java/storyteam/server/story/controller/StoryController.java
@@ -4,34 +4,51 @@ import java.util.List;
 import java.util.Optional;
 
 import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import storyteam.server.story.model.BlocStory;
-import storyteam.server.story.model.CreatorBloc;
 import storyteam.server.story.model.Story;
-import storyteam.server.story.repository.BlocStoryRepository;
+import storyteam.server.story.model.User;
+import storyteam.server.story.services.BlocStoryService;
 import storyteam.server.story.services.StoryService;
+import storyteam.server.story.services.UserService;
 
 @RestController
 @RequestMapping(value = "/story")
 public class StoryController {
-	Logger LOGGER = org.slf4j.LoggerFactory.getLogger(StoryController.class);
-
-	@Autowired
-	BlocStoryRepository BlocStoryRepository;
+	Logger LOGGER = LoggerFactory.getLogger(StoryController.class);
 
 	@Autowired
 	StoryService storyService;
+
+	@Autowired
+	UserService userService;
+
+	@Autowired
+	BlocStoryService blocStoryService;
+
+	@GetMapping(value = "/play/{id}/getFirstBloc")
+	public ResponseEntity<BlocStory> getFirstBlocStory(@PathVariable("id") Integer storyId) {
+		Optional<Story> story = storyService.findById(storyId);
+
+		if (story.isPresent()) {
+			Optional<BlocStory> blocStory = blocStoryService.findById(story.get().getFirstBlocId());
+			if (blocStory.isPresent()) {
+				return ResponseEntity.ok(blocStory.get());
+			}
+		}
+		return ResponseEntity.badRequest().body(null);
+	}
 
 	/**
 	 * Récupère un bloc d'une certaine story
@@ -58,7 +75,7 @@ public class StoryController {
 	 */
 	@GetMapping("/play/{storyId}")
 	ResponseEntity<Story> playStoryStart(@PathVariable("storyId") Integer storyId) {
-		Optional<Story> story = storyService.getStory(storyId);
+		Optional<Story> story = storyService.findById(storyId);
 
 		// Je vais effectuer ici mes test pour récupérer mes futurs commentaires
 		// Attention il faudra prendre garde a ne pas faire de boucle Recup story->
@@ -73,14 +90,22 @@ public class StoryController {
 	}
 
 	/**
-	 * Supprime une story via son id /!\ TODO verif si la story appartient a l'user
-	 * connecté
+	 * Supprime une story via son id
 	 *
 	 * @param storyId
 	 */
 	@DeleteMapping("/delete/{id}")
-	public void deleteStory(@PathVariable("id") Integer storyId) {
-		storyService.delete(storyId);
+	public ResponseEntity<Void> deleteStory(@RequestHeader("Authorization") String auth,
+			@PathVariable("id") Integer storyId) {
+		Optional<User> user = userService.findUserByToken(auth);
+		if (user.isPresent()) {
+			Optional<Story> story = storyService.findById(storyId);
+			if (story.isPresent() && story.get().getUser().equals(user.get())) {
+				storyService.delete(storyId);
+				return ResponseEntity.ok(null);
+			}
+		}
+		return ResponseEntity.noContent().build();
 	}
 
 	/**
@@ -95,7 +120,7 @@ public class StoryController {
 	}
 
 	/**
-	 * Effecture une recherche dans les stories avec les parametres
+	 * Search story via tag selected by the user
 	 *
 	 * @param page
 	 * @param username
@@ -118,11 +143,5 @@ public class StoryController {
 			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 		}
 		return ResponseEntity.ok(storiesSearched);
-	}
-
-	@PostMapping("save")
-	public ResponseEntity<List<BlocStory>> saveNewStory(@RequestBody List<CreatorBloc> blocs) {
-		LOGGER.info("blocs : {}", blocs);
-		return ResponseEntity.ok(storyService.mapCreatorStory(blocs));
 	}
 }
